finalizando serviço de resumo

+++ model/despesa.go
@@ -10,7 +10,7 @@ import (
 type Despesa struct {
 	ID        uint           `gorm:"primary_key"`
 	Descricao string         `gorm:"not null"`
-	Valor     float32        `gorm:"not null"`
+	Valor     float64        `gorm:"not null"`
 	Data      time.Time      `gorm:"not null"`
 	Categoria enum.Categoria `gorm:"not null"`
 }
