+++ model/receita.go
@@ -9,7 +9,7 @@ import (
 type Receita struct {
 	ID        uint      `gorm:"primary_key"`
 	Descricao string    `gorm:"not null"`
-	Valor     float32   `gorm:"not null"`
+	Valor     float64   `gorm:"not null"`
 	Data      time.Time `gorm:"not null"`
 }
